Move string_util.py to cli package
